se hicieron todos los test de logicas faltan los coleccion
Se arreglaron errores de ortografia como perfomer en lugar de performer y se añadieron los controllers y todas las pruebas de logica

+++ src/Album/album.service.ts
@@ -11,34 +11,33 @@ import { promises } from 'dns';
 @Injectable()
 export class AlbumService {
     constructor(
-        @InjectRepository(Albumntity)
+        @InjectRepository(AlbumEntity)
         private albumRepository : Repository<AlbumEntity>,
         @InjectRepository(TrackEntity)
         private track: Repository<TrackEntity>
-
         @InjectRepository(PerformerEntity)
         private performer: Repository<PerformerEntity> 
     ){}
 
     async findall(): Promise<AlbumEntity[]>{
         return await this.albumRepository.find({
-            relations:['track','performer']
+            relations:['tracks','performers']
         });
     }
 
-    async findOne(id: string): Promise<AlbumEntity>{
-        return await this.albumRepository,this.findOne({
-            where: { id},
-            relations: ['track','performer']
+    async findOne(id: string): Promise<AlbumEntity> {
+        const album = await this.albumRepository.findOne({ 
+            where: { id }, 
+            relations: ['tracks', 'performers'] 
         });
-    } 
+    
+        return album;
+    }
 
-    async create(AlbumDto: AlbumDto): Promise<AlbumEntity> {
-        if (new string(AlbumDto.nombre) == " ") {
-            throw new BadRequestException('El nombre no puede estar vacio.');
-        }
-        if (new string(AlbumDto.descripcion) == " ") {
-            throw new BadRequestException('La descripcion no puede estar vacia.');
+    async create(albumDto: AlbumDto): Promise<AlbumEntity> {
+        
+        if (albumDto.descripcion  === '') {
+            throw new BadRequestException('La descripción del álbum no puede estar vacía.');
         }
 
         const album = new AlbumEntity();
@@ -47,7 +46,40 @@ export class AlbumService {
     }
 
     async delete(id: string): Promise<void> {
+        const album = await this.albumRepository.findOne({ 
+            where: { id }, 
+            relations: ['tracks'] 
+        });
+    
+        if (!album) {
+            throw new NotFoundException(`no se encontro el album con el id`);
+        }
+
+        if (album.tracks && album.tracks.length > 0) {
+            throw new BadRequestException('No se puede eliminar un álbum con tracks asociados.');
+        }
         await this.albumRepository.delete(id);
+    }
+
+
+    //Asociacion
+    async addPerformerToAlbum(albumId: string, performerId: string): Promise<AlbumEntity> {
+        const album = await this.albumRepository.findOne({
+            where: { id: albumId },
+            relations: ['performers']
+        });
+        if (!album) throw new NotFoundException('Album no encontrado.');
+    
+        const performer = await this.performer.findOne({
+            where: { id: performerId }
+        });
+        if (!performer) throw new NotFoundException('Performer no encontrado.');
+
+        if (album.performers.length >= 3) {
+            throw new BadRequestException('Un álbum no puede tener más de tres performers asociados.');
+        }
+        album.performers.push(performer);
+        return this.albumRepository.save(album);
     }
